Abstract OauthDataProvider unit tests to share the test code
